% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMSE.R
\name{calc_avg}
\alias{calc_avg}
\title{Calculates the mean of y_val by strategy}
\usage{
calc_avg(dt)
}
\arguments{
\item{dt}{a data table containing y_val that can be grouped by StrategyName}
}
\value{
a datatable of means labelled by StrategyName
}
\description{
Calculates the mean of y_val by strategy
}
