% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMSE.R
\name{calc_prob_below}
\alias{calc_prob_below}
\title{Calculates the proportion of models for each strategy that go below bref}
\usage{
calc_prob_below(dt, bref)
}
\arguments{
\item{dt}{the data table containing some y_values grouped by StrategyName}

\item{bref}{the reference point}
}
\value{
a data.table with the proportions
}
\description{
Calculates the proportion of models for each strategy that go below bref
}
