% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMSE.R
\name{create_table}
\alias{create_table}
\title{Creates a single column of values that can be merged to form a table of MSE values for a species-fleet combination}
\usage{
create_table(results_folder_path, ifunc_groups, n_label_cols, col_name, year,
  first_year, fun, yearly_in_filename = NULL, fleet_group_number = NULL,
  area = 1, scale = 1, ...)
}
\arguments{
\item{results_folder_path}{The path to the folder holding the csv file from which to create a table}

\item{ifunc_groups}{The name of the functional group for which to create the table}

\item{n_label_cols}{The number of label columns in the csv file before the numeric values}

\item{col_name}{The name to give the column being created}

\item{year}{The year in which to calculate the values}

\item{first_year}{The first year of the simulation in the csv file}

\item{yearly_in_filename}{Set to "Yearly" if you want to use a csv file that contains "Yearly" in the name}

\item{fleet_group_number}{The fleet name as specified in the csv file title}

\item{area}{If specified the output values are multiplied by this}

\item{scale}{If specified the output values are divided by this}

\item{bref}{The biomass reference point for which we are calculating the probability of going beneath. Specify in units of t/km2.}
}
\description{
Creates a single column of values that can be merged to form a table of MSE values for a species-fleet combination
}
\examples{
blim_table_2010 <- create_table(results_folder_path = "C:/Users/Mark/Box Sync/Baltic - Stockholm/Results-and-plots-2017-09-08/Results/Biomass/",
fun = calc_prob_below,
ifunc_groups = functional_group,
yearly_in_filename = "Yearly",
n_label_cols = 4,
col_name = 'p(<Blim)',
year = 2010,
first_year = 1990,
bref = 0.4
)

yield_table_2012 <- create_table(results_folder_path = "C:/Users/Mark/Box Sync/Baltic - Stockholm/Results-and-plots-2017-09-08/Results/LandingsTrajectories/",
fun = calc_avg,
ifunc_groups = functional_group,
fleet_group_number = "AllFleets",
yearly_in_filename = "Yearly",
n_label_cols = 5,
col_name = 'avg(Y) (kt)',
year = 2012,
first_year = 2000,
area = 240000,
scale = 1000

)

}
