% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetArea.R
\name{get_area}
\alias{get_area}
\title{Get Area}
\usage{
get_area(path, file.name, area)
}
\arguments{
\item{path}{the path to where the StockAreas.csv file is held}

\item{file.name}{the name of the result file that is being plotted}

\item{area}{the area represented by the functional group found in the result file name}
}
\value{
the area as a numeric(1)
}
\description{
Get Area
}
