% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertFleetName2Number.R
\name{LookupInFile}
\alias{LookupInFile}
\title{LookupInFile}
\usage{
LookupInFile(file, x, header = TRUE)
}
\arguments{
\item{file}{path to the look up file}

\item{x}{the value to look up in the first column}

\item{header}{whether the lookup file contains column headers to skip}
}
\value{
the values in the second column corresponding to the value found in the first column
}
\description{
Looks up x in the first column of file and returns the value in the second column
}
