% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMSE.R
\name{PlotMSE}
\alias{PlotMSE}
\title{Title Plot MSE Plugin Results}
\usage{
PlotMSE(func_group_name = NULL, fleet_group_name = NULL, plot_width_inches,
  plot_height_inches, y_value_label, x_value_label, plot_title = NULL,
  units = "", n_label_cols, yearly_in_filename, results_folder_path,
  save_plot_folder, fleet_lookup_file = NULL, total_area = NULL,
  scale_unit = 1, wrangle_function = NULL, plot_type_fun, ...)
}
\arguments{
\item{func_group_name}{If required, the functional group name to be plotted}

\item{fleet_group_name}{If required, the fleet group name to be plotted}

\item{plot_width_inches}{The number of inches wide for the produced plot}

\item{plot_height_inches}{The number of inches high for the produced plot}

\item{y_value_label}{The y-axis label for the plot}

\item{x_value_label}{The x-axis label for the plot}

\item{plot_title}{Optional The plot title}

\item{units}{The units of the y-axis}

\item{n_label_cols}{The number of columns that label the data before the actual numeric values, contained in the csv being plotted}

\item{yearly_in_filename}{Logical value representing whether the csv file to plot contains in the title the word yearly}

\item{results_folder_path}{The path to the folder that contains the csv file being plotted}

\item{save_plot_folder}{The path to the plots folder which will contain folders for each plot type and inside them the plots created using this function}

\item{fleet_lookup_file}{Contains the path and filename for the fleet lookup file.
The first row contains headers for:
Column1 = Fleet name, Column2 = Fleet Number
This is used when the file names don't contain the fleet name but FleetNoX. It converts from the fleet name to FleetX}

\item{total_area}{The total area covered by the model}

\item{scale_unit}{Used to scale the units, so if you want to convert from tons to kilo tons set to 1000}

\item{wrangle_function}{Set to the type of data manipulation you want depending on the plot you are creating
WrangleBoxplotMean
WrangleBoxplotProbability
WrangleBoxplotAAV
WrangleTimeseriesAnnualChange}

\item{plot_type_fun}{the choice of two types of plots, PlotTimeseries or PlotBoxplot}

\item{...}{If you are using WrangleBoxplotProbability set the argument bref}

\item{true_year_beginning}{The actual date year that is the start of the simulation}
}
\description{
This function contains a complete set of functionality to plot the formal plot types required for an Management Strategy Evaluation Analysis
}
\examples{
PlotMSE(func_group_name = "AdCod",
fleet_group_name = "AllFleets",
start_year = 29, end_year = 29,
true_year_beginning = 1990,
plot_width_inches = 6, plot_height_inches = 4,
y_value_label = "AAV Catch", x_value_label = "Strategy",
units = "\%",
n_label_cols = 5,
yearly_in_filename = TRUE,
results_folder_path = "Results-and-plots-2017-09-08/Results/CatchTrajectories/",
save_plot_folder = "Plots/",
wrangle_function = WrangleBoxplotAAV,
plot_type_fun = PlotBoxplot
)
}
