% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMSE.R
\name{table_default}
\alias{table_default}
\title{Title}
\usage{
table_default(number_decimal_places, functional_group, results_folder_path,
  pdf_path, pdf.height, pdf.width, first_year_hindcast, first_year_forecast,
  first_year_selection, second_year_selection, third_year_selection, blim, bpa,
  area, scale)
}
\arguments{
\item{number_decimal_places}{The number of decimal places to display in the table}

\item{functional_group}{The function group that the table is being created for}

\item{results_folder_path}{The path to the folder holding the results csvs and folders}

\item{pdf_path}{The path to where to save the pdf}

\item{pdf.height}{The pdf height}

\item{pdf.width}{The pdf width}

\item{first_year_hindcast}{The actual first year of the hindcast e.g. 1990}

\item{first_year_forecast}{The actual first year of the forecast e.g. 1990}

\item{first_year_selection}{The actual first year of the selection e.g. 1990}

\item{second_year_selection}{The actual second year of the selection e.g. 1990}

\item{third_year_selection}{The actual third year of the selection e.g. 1990}

\item{blim}{The blim for the given functional group}

\item{bpa}{Bpa for the given functional group}

\item{area}{The area of the EwE model}

\item{scale}{A scaling parameter, e.g. setting to 1000 it will divide the output values by 1000}
}
\description{
Title
}
\examples{
table_default(number_decimal_places = 4, functional_group = "AdCod",
results_folder_path = "C:/Users/Mark/Box Sync/Baltic - Stockholm/Results-and-plots-2017-09-08/Results/",
pdf_path = "C:/Users/Mark/Desktop/Plots/", pdf.height = 2, pdf.width = 17,
first_year_hindcast = 1990, first_year_forecast = 2000, first_year_selection = 2010,
second_year_selection = 2012, third_year_selection = 2015,
blim = 0.4, bpa = 0.5, area = 240000, scale = 1000)
}
