% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMSE.R
\name{WrangleBoxplotAAV}
\alias{WrangleBoxplotAAV}
\title{Calculates the Average Annual Variation in y_val}
\usage{
WrangleBoxplotAAV(dt)
}
\arguments{
\item{dt}{data.table}
}
\value{
data.table
}
\description{
Calculates the Average Annual Variation in y_val
}
