% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMSE.R
\name{WrangleBoxplotMean}
\alias{WrangleBoxplotMean}
\title{Calculates the mean of the data.tables y value grouped by ModelID, StrategyName}
\usage{
WrangleBoxplotMean(dt)
}
\arguments{
\item{dt}{data.table}
}
\value{
data.table
}
\description{
Calculates the mean of the data.tables y value grouped by ModelID, StrategyName
}
