% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMSE.R
\name{WrangleBoxplotProbability}
\alias{WrangleBoxplotProbability}
\title{Calculates the probability of y_val going below bref grouped by Strategy and TimeStep}
\usage{
WrangleBoxplotProbability(dt, bref)
}
\arguments{
\item{dt}{data.table}

\item{bref}{the reference point that is being checked to see if y_val is going beneath}
}
\value{
data.table
}
\description{
Calculates the probability of y_val going below bref grouped by Strategy and TimeStep
}
