% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMSE.R
\name{WrangleTimeseriesAnnualChange}
\alias{WrangleTimeseriesAnnualChange}
\title{Calculates the annual variation of the y_val grouped by TimeStep, ModelID, StrategyName and ResultType}
\usage{
WrangleTimeseriesAnnualChange(dt)
}
\arguments{
\item{dt}{data.table}
}
\value{
data.table
}
\description{
Calculates the annual variation of the y_val grouped by TimeStep, ModelID, StrategyName and ResultType
}
