% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMSE.R
\name{PlotMSE}
\alias{PlotMSE}
\title{Title Plot MSE Plugin Results}
\usage{
PlotMSE(func_group_name = NULL, fleet_group_name = NULL, plot_width_inches,
  plot_height_inches, y_value_label, x_value_label, plot_title = NULL,
  units = "", n_label_cols, yearly_in_filename, results_folder_path,
  save_plot_folder, fleet_lookup_file = NULL, total_area = NULL,
  scale_unit = 1, wrangle_function = NULL, plot_type_fun, ...)
}
\arguments{
\item{func_group_name}{If required, the functional group name to be plotted}

\item{fleet_group_name}{If required, the fleet group name to be plotted}

\item{plot_width_inches}{The number of inches wide for the produced plot}

\item{plot_height_inches}{The number of inches high for the produced plot}

\item{y_value_label}{The y-axis label for the plot}

\item{x_value_label}{The x-axis label for the plot}

\item{plot_title}{Optional The plot title}

\item{units}{The units of the y-axis}

\item{n_label_cols}{The number of columns that label the data before the actual numeric values, contained in the csv being plotted}

\item{yearly_in_filename}{Logical value representing whether the csv file to plot contains in the title the word yearly}

\item{results_folder_path}{The path to the folder that contains the csv file being plotted}

\item{save_plot_folder}{The path to the plots folder which will contain folders for each plot type and inside them the plots created using this function}

\item{fleet_lookup_file}{Contains the path and filename for the fleet lookup file.
The first row contains headers for:
Column1 = Fleet name, Column2 = Fleet Number
This is used when the file names don't contain the fleet name but FleetNoX. It converts from the fleet name to FleetX}

\item{total_area}{The total area covered by the model}

\item{scale_unit}{Used to scale the units, so if you want to convert from tons to kilo tons set to 1000}

\item{wrangle_function}{Set to the type of data manipulation you want depending on the plot you are creating
WrangleBoxplotMean
WrangleBoxplotProbability
WrangleBoxplotAAV
WrangleTimeseriesAnnualChange}

\item{plot_type_fun}{the choice of two types of plots, PlotTimeseries or PlotBoxplot}

\item{...}{If you are using WrangleBoxplotProbability set the argument bref}

\item{true_year_beginning}{The actual date year that is the start of the simulation}
}
\description{
This function contains a complete set of functionality to plot the formal plot types required for an Management Strategy Evaluation Analysis
}
\examples{
plots_folder <- "C:/Users/Mark/Desktop/Plots/"
results_folder <- "C:/Users/Mark/Box Sync/Baltic - Stockholm/Results-and-plots-2017-09-08/Results/"
fleet_lookup_table <- "C:/Users/Mark/Desktop/Plots/FleetNameLookupTable.csv"

#Boxplot-AAV-catch
PlotMSE(
  func_group_name = "AdCod", fleet_group_name = "AllFleets",
  start_year = 1, end_year = 19,
  true_year_beginning = 1990,
  plot_width_inches = 6, plot_height_inches = 4,
  y_value_label = "AAV Catch", x_value_label = "Strategy",
  units = "\%",
  n_label_cols = 5,
  yearly_in_filename = TRUE,
  results_folder_path = paste0(results_folder, "CatchTrajectories/"),
  fleet_lookup_file = fleet_lookup_table,
  save_plot_folder = plots_folder,
  wrangle_function = WrangleBoxplotAAV,
  plot_type_fun = PlotBoxplot
)

#Boxplot-AAV-SSB
PlotMSE(
  func_group_name = "AdCod",
  start_year = 1, end_year = 29,
  true_year_beginning = 1990,
  plot_width_inches = 6, plot_height_inches = 4,
  y_value_label = "AAV SSB", x_value_label = "Years",
  units = "\%",
  n_label_cols = 4,
  yearly_in_filename = TRUE,
  results_folder_path = paste0(results_folder, "Biomass/"),
  save_plot_folder = plots_folder,
  wrangle_function = WrangleBoxplotAAV,
  plot_type_fun = PlotBoxplot
)

#Boxplot-biomass
PlotMSE(
  func_group_name = c("JuvCod"),
  start_year = 1,  end_year = 29,
  true_year_beginning = 1990,
  plot_width_inches = 6, plot_height_inches = 4,
  y_value_label = "Biomass", x_value_label = "Strategy",
  units = "kt",
  n_label_cols = 4,
  yearly_in_filename = TRUE,
  results_folder_path = paste0(results_folder, "Biomass/"),
  save_plot_folder = plots_folder,
  total_area = 240000,
  scale_unit = 1000,
  wrangle_function = WrangleBoxplotMean,
  plot_type_fun = PlotBoxplot
)

#Boxplot-catch
PlotMSE(
  func_group_name = "AdCod", fleet_group_name = "AllFleets",
  start_year = 1,  end_year = 29,
  true_year_beginning = 1990,
  plot_width_inches = 6, plot_height_inches = 4,
  y_value_label = "Average Catch", x_value_label = "Strategy",
  units = "kt/yr",
  n_label_cols = 5,
  yearly_in_filename = TRUE,
  results_folder_path = paste0(results_folder, "CatchTrajectories/"),
  save_plot_folder = plots_folder,
  fleet_lookup_file = fleet_lookup_table,
  total_area = 240000,
  scale_unit = 1000,
  wrangle_function = WrangleBoxplotMean,
  plot_type_fun = PlotBoxplot
)

#Boxplot-effort
PlotMSE(
  fleet_group_name = "ACT0018",
  start_year = 25,  end_year = 29,
  true_year_beginning = 1990,
  plot_width_inches = 6, plot_height_inches = 4,
  y_value_label = "Effort", x_value_label = "Strategy",
  units = "",
  n_label_cols = 4,
  yearly_in_filename = TRUE,
  results_folder_path = paste0(results_folder, "Effort/"),
  save_plot_folder = plots_folder,
  fleet_lookup_file = fleet_lookup_table,
  wrangle_function = WrangleBoxplotMean,
  plot_type_fun = PlotBoxplot
)

#Boxplot-prob-bref
PlotMSE(func_group_name = "AdCod",
        start_year = 24, end_year = 29,
        true_year_beginning = 1990,
        plot_width_inches = 6, plot_height_inches = 4,
        y_value_label = "Probability",  x_value_label = "Strategy",
        units = "",
        n_label_cols = 4,
        yearly_in_filename = TRUE,
        results_folder_path = paste0(results_folder, "Biomass/"),
        save_plot_folder = plots_folder,
        total_area = 240000,
        scale_unit = 1000,
        wrangle_function = WrangleBoxplotProbability,
        plot_type_fun = PlotBoxplot,
        bref = 100, plot_title = "Risk: P(SSB < Blim)"
)

#Boxplot-realisedF
PlotMSE(func_group_name = "AdCod",
        start_year = 0,  end_year = 29,
        true_year_beginning = 1990,
        plot_width_inches = 6, plot_height_inches = 4,
        y_value_label = "Realised F",  x_value_label = "Strategy",
        units = "/km2/yr",
        n_label_cols = 5,
        yearly_in_filename = TRUE,
        results_folder_path = paste0(results_folder, "RealisedF/"),
        save_plot_folder = plots_folder,
        wrangle_function = WrangleBoxplotMean,
        plot_type_fun = PlotBoxplot
)

#Boxplot-TAC
PlotMSE(func_group_name = "AdCod", fleet_group_name = "AllFleets",
        start_year = 5, end_year = 29,
        true_year_beginning = 1990,
        plot_width_inches = 6, plot_height_inches = 4,
        y_value_label = "TAC",  x_value_label = "Strategy",
        units = "kt",
        n_label_cols = 5,
        yearly_in_filename = FALSE,
        results_folder_path = paste0(results_folder, "HCRQuota_Targ/"),
        save_plot_folder = plots_folder,
        fleet_lookup_file = fleet_lookup_table,
        total_area = 240000,
        scale_unit = 1000,
        wrangle_function = WrangleBoxplotMean,
        plot_type_fun = PlotBoxplot
)

#Boxplot-effort
PlotMSE(
  fleet_group_name = "ACT0018",
  start_year = 25,  end_year = 29,
  true_year_beginning = 1990,
  plot_width_inches = 6, plot_height_inches = 4,
  y_value_label = "Effort", x_value_label = "Strategy",
  units = "",
  n_label_cols = 4,
  yearly_in_filename = TRUE,
  results_folder_path = paste0(results_folder, "Effort/"),
  fleet_lookup_file = fleet_lookup_table,
  save_plot_folder = plots_folder,
  fleet_lookup_file = fleet_lookup_table,
  wrangle_function = WrangleBoxplotMean,
  plot_type_fun = PlotBoxplot
)


#Timeseries-biomass
PlotMSE(
  func_group_name = "AdCod",
  true_year_beginning = 1990,
  plot_width_inches = 12, plot_height_inches = 3,
  y_value_label = "Biomass", x_value_label = "Years",
  units = "kt",
  n_label_cols = 4,
  yearly_in_filename = TRUE,
  results_folder_path = paste0(results_folder, "Biomass/"),
  save_plot_folder = plots_folder,
  total_area = 240000,
  scale_unit = 1000,
  plot_type_fun = PlotTimeseries
)

#Timeseries-realisedF
PlotMSE(func_group_name = "AdCod",
        start_year = 4, end_year = 29,
        true_year_beginning = 1990,
        plot_width_inches = 12, plot_height_inches = 3,
        y_value_label = "Realised F",  x_value_label = "Years",
        units = "/km2/yr",
        n_label_cols = 5,
        yearly_in_filename = TRUE,
        results_folder_path = paste0(results_folder, "RealisedF/"),
        save_plot_folder = plots_folder,
        plot_type_fun = PlotTimeseries
)

#Timeseries-annual-change-TAC
PlotMSE(
  func_group_name = "AdCod", fleet_group_name = "AllFleets",
  true_year_beginning = 1990,
  plot_width_inches = 12, plot_height_inches = 3,
  y_value_label = "Annual Change in TAC", x_value_label = "Years",
  units = "\%",
  n_label_cols = 5,
  yearly_in_filename = FALSE,
  results_folder_path = paste0(results_folder, "HCRQuota_Targ/"),
  save_plot_folder = plots_folder,
  fleet_lookup_file = fleet_lookup_table,
  wrangle_function = WrangleTimeseriesAnnualChange,
  plot_type_fun = PlotTimeseries
)

#Timeseries-catch
PlotMSE(
  func_group_name = "AdCod", fleet_group_name = "ACT0018",
  true_year_beginning = 1990,
  plot_width_inches = 12, plot_height_inches = 3,
  y_value_label = "Catch", x_value_label = "Years",
  units = "kt/yr",
  n_label_cols = 5,
  yearly_in_filename = TRUE,
  results_folder_path = paste0(results_folder, "CatchTrajectories/"),
  save_plot_folder = plots_folder,
  fleet_lookup_file = fleet_lookup_table,
  total_area = 240000,
  scale_unit = 1000,
  plot_type_fun = PlotTimeseries
)

#Timeseries-effort
PlotMSE(
  fleet_group_name = "ACT0018",
  start_year = NULL, end_year = 29,
  true_year_beginning = 2004,
  plot_width_inches = 12, plot_height_inches = 3,
  y_value_label = "Effort", x_value_label = "Years",
  n_label_cols = 4,
  yearly_in_filename = TRUE,
  results_folder_path = paste0(results_folder, "Effort/"),
  save_plot_folder = plots_folder,
  fleet_lookup_file = fleet_lookup_table,
  plot_type_fun = PlotTimeseries
)

#Timeseries-TAC
PlotMSE(
  func_group_name = "AdCod", fleet_group_name = "ACT1824",
  true_year_beginning = 1990,
  plot_width_inches = 12, plot_height_inches = 3,
  y_value_label = "TAC", x_value_label = "Years",
  units = "kt",
  n_label_cols = 5,
  yearly_in_filename = FALSE,
  results_folder_path = paste0(results_folder, "HCRQuota_Targ/"),
  save_plot_folder = plots_folder,
  fleet_lookup_file = fleet_lookup_table,
  total_area = 240000,
  scale_unit = 1000,
  plot_type_fun = PlotTimeseries
)
}
