% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMSE.R
\name{save_table_pdf}
\alias{save_table_pdf}
\title{Saves the table to}
\usage{
save_table_pdf(list_of_tables, name_of_file, pdf_path, pdf.height, pdf.width)
}
\arguments{
\item{pdf.width}{}
}
\description{
Saves the table to
}
